Generalize update details component
